Refactor TodoList example to use mapDispatchToProps
